+++ CHANGELOG.md
@@ -7,6 +7,10 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 - **[UPDATE]** Updated to VisUI 1.4.10.
 - **[FEATURE]** (`ktx-app`) `clearScreen` now accepts additional `clearDepth` boolean parameter that controls whether 
 the `GL_DEPTH_BUFFER_BIT` is added to the mask.
+- **[FEATURE]** (`ktx-assets-async`) Added `AssetStorageSnapshot` class that stores a copy of `AssetStorage` state
+for debugging purposes. Supports formatted string output with `prettyFormat`.
+- **[FEATURE]** (`ktx-assets-async`) `AssetStorage` now includes `takeSnapshot` and `takeSnapshotAsync` methods that
+allow to copy and inspect the internal state of the storage for debugging purposes.
 
 #### 1.9.13-b1